delete all orphan addresses

+++ src/main/java/ru/ssk/service/impl/AddressServiceImpl.java
@@ -51,12 +51,16 @@ public class AddressServiceImpl implements AddressService {
     @Override
     public void deleteIfOrphan(Address address) {
         address = addressRepository.findOne(address.getId());
-        System.out.println("Persons : " + address.getPersons().size());
-        System.out.println("Passports : " + address.getPassports().size());
-        if (address.getPersons().isEmpty() && address.getPassports().isEmpty()) {
+        if (address != null && address.getPersons().isEmpty() && address.getPassports().isEmpty()) {
             addressRepository.delete(address);
             addressRepository.flush();
         }
+    }
+
+    @Override
+    public void deleteOrphans() {
+        addressRepository.deleteOrphans();
+        addressRepository.flush();
     }
 
     @Override
